+++ CodePub/multigrid.h
@@ -1,7 +1,7 @@
-double v_cycle(int max_recursion, int c, int n, int m, double L,
+int v_cycle(int max_recursion, int c, int n, int m, double L,
                                      int **ia_ptr, int **ja_ptr, double **a_ptr, double *b, double *x, void *Numeric);
 
-double w_cycle(int max_recursion, int c, int n, int m, double L,
+int w_cycle(int max_recursion, int c, int n, int m, double L,
                                      int **ia_ptr, int **ja_ptr, double **a_ptr, double *b, double *x, void *Numeric);
 
 void *generate_multigrid_problem(int max_recursion, int m, int **ia_ptr, int **ja_ptr, double **a_ptr,
